gateway/src/main/webapp/app/entities/user-auth/route/user-auth-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { IUserAuth, UserAuth } from '../user-auth.model';
import { UserAuthService } from '../service/user-auth.service';

import { UserAuthRoutingResolveService } from './user-auth-routing-resolve.service';

describe('UserAuth routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: UserAuthRoutingResolveService;
  let service: UserAuthService;
  let resultUserAuth: IUserAuth | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(UserAuthRoutingResolveService);
    service = TestBed.inject(UserAuthService);
    resultUserAuth = undefined;
  });

  describe('resolve', () => {
    it('should return IUserAuth returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultUserAuth = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultUserAuth).toEqual({ id: 123 });
    });

    it('should return new IUserAuth if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultUserAuth = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultUserAuth).toEqual(new UserAuth());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as UserAuth })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultUserAuth = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultUserAuth).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
